+++ tasks/register/compileAssets.js
@@ -1,9 +1,9 @@
 module.exports = function (grunt) {
-	grunt.registerTask('compileAssets', [
-		'clean:dev',
-		'jst:dev',
-		'less:dev',
-		'copy:dev',
-		'coffee:dev'
-	]);
+  grunt.registerTask('compileAssets', [
+    'clean:dev',
+    'handlebars:dev',       // changed jst task to handlebars task
+    'less:dev',
+    'copy:dev',
+    'coffee:dev'
+  ]);
 };